+++ CraftIQ.Inventory.API/CraftIQ.Inventory.API.csproj
@@ -9,6 +9,21 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <Compile Remove="Endpoints\Categories\Deleting\**" />
+    <Compile Remove="Endpoints\Categories\Reading\**" />
+    <Compile Remove="Endpoints\Categories\Updating\**" />
+    <Content Remove="Endpoints\Categories\Deleting\**" />
+    <Content Remove="Endpoints\Categories\Reading\**" />
+    <Content Remove="Endpoints\Categories\Updating\**" />
+    <EmbeddedResource Remove="Endpoints\Categories\Deleting\**" />
+    <EmbeddedResource Remove="Endpoints\Categories\Reading\**" />
+    <EmbeddedResource Remove="Endpoints\Categories\Updating\**" />
+    <None Remove="Endpoints\Categories\Deleting\**" />
+    <None Remove="Endpoints\Categories\Reading\**" />
+    <None Remove="Endpoints\Categories\Updating\**" />
+  </ItemGroup>
+
+  <ItemGroup>
     <PackageReference Include="huzcodes.Endpoints" Version="1.1.0" />
     <PackageReference Include="huzcodes.Extensions" Version="1.2.0" />
     <PackageReference Include="Microsoft.EntityFrameworkCore" Version="8.0.3" />
@@ -22,6 +37,12 @@
 
   <ItemGroup>
     <ProjectReference Include="..\CraftIQ.Inventory.Infrastructure\CraftIQ.Inventory.Infrastructure.csproj" />
+    <ProjectReference Include="..\CraftIQ.Inventory.Services\CraftIQ.Inventory.Services.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Folder Include="Endpoints\Categories\Update\" />
+    <Folder Include="Endpoints\Categories\Read\" />
   </ItemGroup>
 
 </Project>
