+++ CraftIQ.Inventory.API/Endpoints/Categories/Create/Categories.CreateCategoriesResponse.cs
@@ -1,4 +1,4 @@
-﻿namespace CraftIQ.Inventory.API.Endpoints.Categories
+﻿namespace CraftIQ.Inventory.API.Endpoints.Categories.Create
 {
     public class CreateCategoriesResponse
     {
