+++ CraftIQ.Inventory.API/Endpoints/Categories/Create/Categories.cs
@@ -1,26 +1,26 @@
-﻿using CraftIQ.Inventory.Core.Entities;
+﻿using CraftIQ.Inventory.Core.Interfaces;
+using CraftIQ.Inventory.Shared.Contracts.Categories;
 using huzcodes.Endpoints.Abstractions;
 using huzcodes.Extensions.Exceptions;
-using huzcodes.Persistence.Interfaces.Repositories;
 using Microsoft.AspNetCore.Mvc;
 
-namespace CraftIQ.Inventory.API.Endpoints.Categories
+namespace CraftIQ.Inventory.API.Endpoints.Categories.Create
 {
-    public class Categories(IRepository<Category> repository) : EndpointsAsync
+    public class Categories(ICategoriesServices services) : EndpointsAsync
                                                                .WithRequest<CreateCategoriesRequest>
                                                                .WithActionResult<CreateCategoriesResponse>
     {
-        private readonly IRepository<Category> _repository = repository;
+        private readonly ICategoriesServices _services = services;
 
-        [HttpPost(Routes.CategoriesRoutes.Create)]
+        [HttpPost(Routes.CategoriesRoutes.BaseUrl)]
         public override async Task<ActionResult<CreateCategoriesResponse>> HandleAsync
                        (CreateCategoriesRequest request, CancellationToken cancellationToken = default)
         {
             if (request == null)
                 throw new ResultException("request can't be null", StatusCodes.Status400BadRequest);
 
-            var oData = new Category(request.Name, request.Description);
-            var oResult = await _repository.AddAsync(oData);
+            var oData = new CategoriesOperationsContract(request.Name, request.Description);
+            var oResult = await _services.CreateCategory(oData);
             return Ok(new CreateCategoriesResponse(oResult.Name, oResult.Description));
         }
     }
