+++ CraftIQ.Inventory.API/Endpoints/Categories/Delete/Categories.DeleteRequest.cs
@@ -1,0 +1,10 @@
+﻿using Microsoft.AspNetCore.Mvc;
+
+namespace CraftIQ.Inventory.API.Endpoints.Categories.Delete
+{
+    public class CategoriesDeleteRequest
+    {
+        [FromRoute]
+        public Guid categoryId { get; set; }
+    }
+}
