+++ CraftIQ.Inventory.API/Routes.cs
@@ -4,7 +4,23 @@
     {
         public class CategoriesRoutes
         {
-            public const string Create = "/category";
+            public const string BaseUrl = "/category";
+            public const string Delete = BaseUrl + "{categoryId}";
+
+            // Cars
+            // POST baseUrl/Cars
+            // GET baseUrl/Cars
+            // GET BY ID baseUrl/Cars/{id}
+            // UPDATE baseUrl/Cars/{id}
+            // DELETE baseUrl/Cars/{id}
+
+            // Engines
+            //POST baseUrl/Cars/Engines
+            // PUT baseUrl/Cars/Engines/{Id}
+            // DELETE baseUrl/Cars/Engines/{Id}
+            // GET baseUrl/Cars/Engines
+            // GET baseUrl/Cars/{id}/Engines
+            // GET baseUrl/Cars/{id}/Engines/{Id}
         }
     }
 }
